Add group A transaction type INDV_REC screen
