+++ front-end/src/app/shared/models/transaction-types/INDV_REC.model.ts
@@ -1,9 +1,9 @@
 import { TransactionType } from '../../interfaces/transaction-type.interface';
-import { schema } from 'fecfile-validate/fecfile_validate_js/dist/OFFSET_TO_OPEX';
 import { SchATransaction } from '../scha-transaction.model';
 import { ScheduleATransactionTypes } from '../scha-transaction.model';
 import { ScheduleATransactionTypeLabels } from '../scha-transaction.model';
 import { LabelUtils } from 'app/shared/utils/label.utils';
+import { schema } from 'fecfile-validate/fecfile_validate_js/dist/INDV_REC';
 
 export class INDV_REC implements TransactionType {
   scheduleId = 'A';
