+++ front-end/src/app/shared/models/transaction-types/OFFSET_TO_OPEX.model.ts
@@ -1,9 +1,9 @@
 import { TransactionType } from '../../interfaces/transaction-type.interface';
-import { schema } from 'fecfile-validate/fecfile_validate_js/dist/OFFSET_TO_OPEX';
 import { SchATransaction } from '../scha-transaction.model';
 import { ScheduleATransactionTypes } from '../scha-transaction.model';
 import { ScheduleATransactionTypeLabels } from '../scha-transaction.model';
 import { LabelUtils } from 'app/shared/utils/label.utils';
+import { schema } from 'fecfile-validate/fecfile_validate_js/dist/OFFSET_TO_OPEX';
 
 export class OFFSET_TO_OPEX implements TransactionType {
   scheduleId = 'A';
