+++ front-end/src/app/transactions/transaction-group-a/transaction-group-a.component.html
@@ -15,6 +15,7 @@
             optionLabel="name"
             optionValue="code"
             [readonly]="true"
+            class="readonly"
           ></p-dropdown>
         </div>
       </div>
@@ -151,7 +152,7 @@
     <p-divider></p-divider>
     <h3>Employer</h3>
     <div class="grid">
-      <div class="col-12">
+      <div class="col-6">
         <div class="field">
           <label for="contributor_employer">EMPLOYER</label>
           <input type="text" pInputText id="contributor_employer" formControlName="contributor_employer" />
@@ -162,9 +163,7 @@
           ></app-error-messages>
         </div>
       </div>
-    </div>
-    <div class="grid">
-      <div class="col-12">
+      <div class="col-6">
         <div class="field">
           <label for="contributor_occupation">OCCUPATION</label>
           <input type="text" pInputText id="contributor_occupation" formControlName="contributor_occupation" />
@@ -217,7 +216,13 @@
       <div class="col-6">
         <div class="field">
           <label for="contribution_amount">CONTRIBUTION AMOUNT</label>
-          <input type="text" pInputText id="contribution_amount" formControlName="contribution_amount" />
+          <p-inputNumber
+            inputId="contribution_amount"
+            formControlName="contribution_amount"
+            mode="currency"
+            currency="USD"
+            locale="en-US"
+          ></p-inputNumber>
           <app-error-messages
             [form]="form"
             fieldName="contribution_amount"
@@ -230,14 +235,15 @@
       <div class="col-6">
         <div class="field">
           <label for="contribution_aggregate">CONTRIBUTION AGGREGATE</label>
-          <input
-            type="text"
-            pInputText
-            id="contribution_aggregate"
+          <p-inputNumber
+            inputId="contribution_aggregate"
             formControlName="contribution_aggregate"
-            class="readonly"
-            readonly
-          />
+            mode="currency"
+            currency="USD"
+            locale="en-US"
+            [readonly]="true"
+            inputStyleClass="readonly"
+          ></p-inputNumber>
           <app-error-messages
             [form]="form"
             fieldName="contribution_aggregate"
@@ -261,6 +267,8 @@
             pInputTextarea
             [autoResize]="true"
             formControlName="contribution_purpose_descrip"
+            class="readonly"
+            readonly
           ></textarea>
           <app-error-messages
             [form]="form"
