+++ front-end/src/app/transactions/transaction-group-b/transaction-group-b.component.html
@@ -211,7 +211,7 @@
         <div class="field">
           <label for="contribution_amount">CONTRIBUTION AMOUNT</label>
           <p-inputNumber
-            id="contribution_amount"
+            inputId="contribution_amount"
             formControlName="contribution_amount"
             mode="currency"
             currency="USD"
@@ -230,13 +230,13 @@
         <div class="field">
           <label for="contribution_aggregate">CONTRIBUTION AGGREGATE</label>
           <p-inputNumber
-            id="contribution_aggregate"
+            inputId="contribution_aggregate"
             formControlName="contribution_aggregate"
             mode="currency"
             currency="USD"
             locale="en-US"
             [readonly]="true"
-            styleClass="readonly"
+            inputStyleClass="readonly"
           ></p-inputNumber>
           <app-error-messages
             [form]="form"
